// src/app/model/series-details.ts
export interface SeriesDetails {
  avg_run_time: number;
  backdrop_path: string;
  episode_run_time: any;
  first_air_date: Date;
  last_air_date: Date;
  last_episode_to_air: Date;
  homepage: string;
  id: number;
  original_lan: string;
  networks: any;
  no_of_seasons: number;
  no_of_episodes: number;
  original_title: string;
  overview: string;
  poster_path: string;
  production_companies: any;
  production_countries: any;
  popularity: number;
  tagline: string;
  status: string;
  vote_average: number;
  vote_count: number;
  seasons: any;
  genre: any;
  watchprovider: any;
  background_image: any;

  backdropList: any;
  logoList: any;
  posterList: any;
  reviewList: any;
  castList: any;
  crewList: any;
  similartvshowList: any;
  rectvshowList: any;
  videoList: any;

  title: string;
  release_date: string;
  job: string;
}